requirements.txt
cairosvg==2.5.2
Pillow==9.1.1
